+++ lib/heroku-resque-workers-scaler/config.rb
@@ -1,42 +1,51 @@
+require 'yaml'
+
 module HerokuResqueAutoScale
   module Config
     extend self
-    
+
     CONFIG_FILE_NAME = 'scaler_config.yml'
-    
+
+    def mode
+      @mode ||= config['mode']
+    end
+
+    def threshold
+      @threshold ||= config['threshold']
+    end
+
     def thresholds
-      @thresholds ||= begin
-        if config_file?
-          config['thresholds']
-        else
-          [{workers:1,job_count:1},{workers:2,job_count:15},{workers:3,job_count:25},{workers:4,job_count:40},{workers:5,job_count:60}]
-        end
-      end
+      @thresholds ||= config['thresholds']
     end
-    
+
     def environments
-      @environments ||= begin
-        if config_file?
-          config['environments']
-        else
-          [ 'production' ]
-        end
-      end
+      @environments ||= config['environments']
     end
-    
-    private 
-    
-    def config_file?
-      @config_file ||= override? || File.exists?(CONFIG_FILE_NAME)
+
+    def worker_name
+      @worker_name ||= config['worker_name']
     end
-    
+
+    def worker_count_key
+      @worker_count_key ||= config['worker_count_key']
+    end
+
+    def heroku_cache_duration
+      @heroku_cache_duration ||= config['heroku_cache_duration']
+    end
+
+    private
+
     def config
-      @config ||= override? ? YAML.load_file(Rails.root.join("config/#{CONFIG_FILE_NAME}").to_s) : YAML.load_file(CONFIG_FILE_NAME)
+      @config ||= YAML.load_file(config_route_path)
     end
-    
-    def override?
-      File.exists?(Rails.root.join("config/#{CONFIG_FILE_NAME}").to_s) rescue false
+
+    def config_route_path
+      if defined?(Rails) && File.exists?(Rails.root.join("config/#{CONFIG_FILE_NAME}").to_s)
+        Rails.root.join("config/#{CONFIG_FILE_NAME}").to_s
+      else
+        File.expand_path("../../../config/#{CONFIG_FILE_NAME}", __FILE__)
+      end
     end
   end
 end
-